+++ lib/application/impl/task/task_delete_service_impl.dart
@@ -1,3 +1,4 @@
+import 'package:nova_chrono/application/api/exception/task_not_found_exception.dart';
 import 'package:nova_chrono/application/api/task/task_delete_service.dart';
 
 import '../../../domain/repository/task_repository.dart';
@@ -12,7 +13,12 @@ class TaskDeleteServiceImpl implements TaskDeleteService {
 
   @override
   Future<void> deleteTask(String taskId) async {
-    // TODO: Check if task exists
+    var task = await _taskRepository.getById(taskId);
+
+    if (task == null) {
+      throw TaskNotFoundException('Task with id $taskId not found.');
+    }
+
     await _taskRepository.deleteTask(taskId);
   }
 }
